addStudentToSubject moved from studentController to SubjectController
