+++ app/Http/Controllers/StudentController.php
@@ -139,27 +139,7 @@ class StudentController extends Controller
         $semester->students()->attach($request->student_id);
     }
 
-    /**
-     * @param int id (subject id)
-     * this function add a students to a particular subject
-     */
-    public function addStudentToSubject(Request $request)
-    {
-        $this->validate($request, [
-            'student_id' => 'required|integer',
-            'subject_id' => 'required|integer'
-        ]);
-
-        $subject = Subject::findOrFail($request->subject_id);
-        $subject->students()->attach($request->student_id);
-        
-        // if a student is added to a subject, also add them to the related semester automatically
-        $semester = $subject->semester;
-        $student = $semester->students->where('id', $request->student_id)->first();
-        if($student == null){
-            $semester->students()->attach($request->student_id);
-        }
-    }
+    
 
     // the update function never updates user_id as it is only set once!
     public function update(Request $request, $id)
